update the event document class and routes

+++ planner/routes/events.py
@@ -30,6 +30,7 @@ async def retrieve_event(id: PydanticObjectId) -> Event:
 
 @event_router.post("/new")
 async def create_event(body: Event, user: str = Depends(authenticate)) -> Event:
+    body.creator = user
     await event_database.save(body)
     return {
         "message": "Event created successfully",
@@ -40,24 +41,36 @@ async def create_event(body: Event, user: str = Depends(authenticate)) -> Event:
 async def update_event(
     id: PydanticObjectId, body: EventUpdate, user: str = Depends(authenticate)
 ) -> Event:
+    event = await event_database.update(id, body)
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_400_BAD_REQUEST,
+            detail="Operation not allowed",
+        )
     updated_event = await event_database.update(id, body)
     if not updated_event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
             detail="Event with supplied ID does not exist",
         )
-
     return updated_event
 
 
 @event_router.delete("/{id}")
 async def delete_event(id: PydanticObjectId, user: str = Depends(authenticate)) -> dict:
-    event = await event_database.delete(id)
+    event = await event_database.get(id)
     if not event:
         raise HTTPException(
             status_code=status.HTTP_404_NOT_FOUND,
-            detail="Event with supplied ID does not exist",
+            detail="Event not found",
         )
-        return {
-            "message": "Event deleted successfully",
-        }
+    if event.creator != user:
+        raise HTTPException(
+            status_code=status.HTTP_404_NOT_FOUND,
+            detail="Operation not allowed",
+        )
+    event = await event_database.delete(id)
+
+    return {
+        "message": "Event deleted successfully",
+    }
